refactor: Update allowedOrigin variable in backend/index.js
This commit updates the allowedOrigin variable in the backend/index.js file. The condition for setting the allowedOrigin value has been modified to use the NODE_ENV environment variable. If the NODE_ENV is set to 'product', the allowedOrigin will be set to prodOrigins, otherwise it will be set to devOrigins. This change ensures that the correct allowed origins are used for CORS configuration based on the environment.